pyproject.toml
[project]
name = "voiceadmin-learn"
version = "0.1.0"
description = "Hybrid voice agent for eligibility verification — learning project."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic==0.97.0",
    "cartesia==3.0.2",
    "deepgram-sdk==6.1.1",
    "elevenlabs>=2.45.0",
    "fastapi==0.136.0",
    "groq==1.1.2",
    "langfuse==4.5.0",
    "pipecat-ai==1.0.0",
    "python-dotenv==1.2.2",
    "python-multipart==0.0.26",
    "structlog==25.5.0",
    "twilio==9.10.5",
    "uvicorn==0.44.0",
]

[dependency-groups]
dev = [
    "hypothesis==6.152.4",
    "pre-commit==4.5.1",
    "pyright==1.1.408",
    "pytest==9.0.3",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.0.0",
    "ruff==0.15.11",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "W", "SIM", "RUF"]
ignore = ["E501"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
strict = ["agent", "tests"]
include = ["agent", "tests"]
exclude = [".venv", "**/__pycache__"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
# `--cov-fail-under` enforces the floor in CI; `--no-cov-on-fail` keeps the
# report visible when tests fail so we can debug without hunting for the
# coverage section.
addopts = "-q --cov=agent --cov-report=term-missing --cov-fail-under=90 --no-cov-on-fail"

[tool.coverage.run]
source = ["agent"]
# `agent/main.py` is the FastAPI + Pipecat entrypoint — exercised by live
# phone calls (M7'), not unit tests. Including it would tank the coverage
# floor for code that's intentionally not unit-tested.
omit = ["agent/main.py"]

[tool.coverage.report]
# Fail on missing branches in covered files (skip omitted ones).
show_missing = true
skip_empty = true
# Per-file thresholds aren't built into coverage; the global floor + the
# per-file `Missing` columns in `term-missing` make gaps visible enough
# for review. Tighten via PR review when a file drops below ~85%.
